Conversational AI Data Science Tutor/requirements.txt
streamlit
langchain
langchain-google-genai
google-auth
google-generativeai
spacy
